Rename config.[ch] to cfg.[ch]
Renames `config.[ch]` to `cfg.[ch]` so we can use the name `config.h`
for generated header file for the whole project.
