+++ src/agent/agent.h
@@ -5,8 +5,8 @@
 #include <stdbool.h>
 #include <sys/socket.h>
 
+#include "libhirte/common/cfg.h"
 #include "libhirte/common/common.h"
-#include "libhirte/common/config.h"
 #include "libhirte/hashmap/hashmap.h"
 
 typedef struct Agent Agent;
