+++ src/libhirte/meson.build
@@ -5,9 +5,9 @@ libhirte_src = [
     'bus/bus.h',
     'bus/utils.c',
     'bus/utils.h',
+    'common/cfg.c',
+    'common/cfg.h',
     'common/common.h',
-    'common/config.c',
-    'common/config.h',
     'common/event-util.c',
     'common/event-util.h',
     'common/protocol.c',
